pr fixes
- mainly moved lib versions to version catalog
